Refactor: AlbumList 컴포넌트 파일분리
